// src/app/core/models/member.model.ts
export interface Member {
  id: string;
  firstName: string;
  lastName: string;
  fullName: string;
  designation: string;
  phone: string;
  email: string;
  gender: string;
  dob: string;
  address: string;
  planName: string;
  startDate: string;
  duration: number;
  trainer: string;
  emergencyContact: string;
  photo: string;
  govId: string;
  notes: string;
  active: boolean;
}